Fix bug causing transition selection to drop
Editing a transition via the Inspector pallete would drop selection
if the transition was not the only child of the parent state, due to
a bug in the code looking for the matching transition.

After fixing the bug, tried to make the heuristic a little more robust
with extra fallback criteria.