Add `Reset Visualization` command
Note that this does not remove the `xmlns:viz` attribute from the root.
It should do that, but the visualization editor will add it again.
To do this completely would probably need the command
to disconnect and close the visual editor and dispose the glue
right after the visual editor resets the document.

Fixes #28